pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prow-ci-mcp-server"
version = "0.1.0"
description = "Prow CI MCP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.10.6",
    "httpx>=0.28.1",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "python-dateutil>=2.9.0.post0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[tool.hatch.build.targets.wheel]
only-include = ["mcp_server/", "main.py"]
